requirements.txt
streamlit>=1.28.0
yfinance>=0.2.28
pandas>=2.0.0
pandas-ta>=0.3.14b0
plotly>=5.17.0
kaleido>=0.2.1
qrcode[pil]>=7.4.2
requests>=2.31.0
python-dotenv>=1.0.0
flask>=2.3.0
pillow>=10.0.0
